Moved private setter class in package org.lobobrowser.html.style.setter
