+++ HTML_Renderer/src/main/java/org/lobobrowser/html/style/setter/Property.java
@@ -1,0 +1,23 @@
+package org.lobobrowser.html.style.setter;
+
+public class Property {
+
+	/** The value. */
+	public final String value;
+
+	/** The important. */
+	public final boolean important;
+
+	/**
+	 * Instantiates a new property.
+	 *
+	 * @param value
+	 *            the value
+	 * @param important
+	 *            the important
+	 */
+	public Property(final String value, final boolean important) {
+		this.value = value;
+		this.important = important;
+	}
+}
